feat: improve settings flow and refine home page UI
