+++ src/app/[locale]/auth/ui/index.ts
@@ -1,2 +1,1 @@
 export * from "./language-toggle";
-export * from "./theme-toggle";
