--- src/app/providers/theme-provider.tsx
@@ -1,84 +1,0 @@
-"use client";
-
-import type React from "react";
-import type { PropsWithChildren } from "react";
-import { createContext, useCallback, useContext, useEffect, useMemo, useState } from "react";
-
-type Theme = "light" | "dark";
-
-type ThemeContextValue = {
-  theme: Theme;
-  toggleTheme: () => void;
-  setTheme: (theme: Theme) => void;
-  mounted: boolean;
-};
-
-const ThemeContext = createContext<ThemeContextValue | null>(null);
-
-function getSystemTheme(): Theme {
-  if (typeof window === "undefined") {
-    return "dark";
-  }
-
-  return window.matchMedia("(prefers-color-scheme: dark)").matches ? "dark" : "light";
-}
-
-function applyTheme(theme: Theme) {
-  const root = document.documentElement;
-
-  root.setAttribute("data-theme", theme);
-  root.style.colorScheme = theme;
-}
-
-export function ThemeProvider({ children }: PropsWithChildren): React.JSX.Element {
-  const [theme, setThemeState] = useState<Theme>("dark");
-  const [mounted, setMounted] = useState(false);
-
-  useEffect(() => {
-    const stored = window.localStorage.getItem("theme") as Theme | null;
-    const nextTheme = stored ?? getSystemTheme();
-
-    setThemeState(nextTheme);
-    applyTheme(nextTheme);
-    setMounted(true);
-  }, []);
-
-  const setTheme = useCallback((nextTheme: Theme) => {
-    setThemeState(nextTheme);
-    window.localStorage.setItem("theme", nextTheme);
-    applyTheme(nextTheme);
-  }, []);
-
-  const toggleTheme = useCallback(() => {
-    setThemeState((prevTheme) => {
-      const nextTheme = prevTheme === "dark" ? "light" : "dark";
-
-      window.localStorage.setItem("theme", nextTheme);
-      applyTheme(nextTheme);
-
-      return nextTheme;
-    });
-  }, []);
-
-  const value = useMemo<ThemeContextValue>(
-    () => ({
-      theme,
-      toggleTheme,
-      setTheme,
-      mounted,
-    }),
-    [mounted, setTheme, theme, toggleTheme],
-  );
-
-  return <ThemeContext.Provider value={value}>{children}</ThemeContext.Provider>;
-}
-
-export function useTheme(): ThemeContextValue {
-  const context = useContext(ThemeContext);
-
-  if (!context) {
-    throw new Error("useTheme must be used inside ThemeProvider");
-  }
-
-  return context;
-}
