+++ src/shared/lib/theme/theme-provider.tsx
@@ -1,7 +1,6 @@
 "use client";
 
 import React, { createContext, useCallback, useContext, useEffect, useMemo, useState } from "react";
-
 import type { PropsWithChildren } from "react";
 
 export type Theme = "light" | "dark";
