+++ src/shared/ui/icon/Icon.tsx
@@ -1,8 +1,8 @@
 "use client";
+
 import React from "react";
 
 import AntdIcon from "@ant-design/icons";
-
 import {
   SettingOutlined,
   HomeOutlined,
@@ -11,6 +11,7 @@ import {
   CrownOutlined,
   CheckOutlined,
 } from "@ant-design/icons";
+
 import { cn } from "../../lib";
 
 type IconType = "settings" | "home" | "practice" | "glossary" | "rating" | "check";
